+++ run.sh
@@ -3,7 +3,7 @@
 #git pull
 rm -f out.txt
 #python3 -m main 2 designs/test-designs/test_2.v --sv -t place_holder_2
-python3 -m main 2 designs/test-designs/test_2.v --sv 
+python3 -m main 3 designs/test-designs/test_2.v --sv 
 #python3 -m main 2 designs/aes/aes.F --sv 
 #python3 -m main 2 designs/test-designs/updowncounter.v --sv 
 #python3 -m main 1 filelist.F --sv